// a3front/src/modules/negotiations/supplier-new/enums/supplier-progress-modules.enum.ts
export enum SupplierProgressModulesEnum {
  CLASSIFICATIONS = 18, // Clasificaciones
  GENERAL_INFORMATION = 1, // Información general
  COMMERCIAL_LOCATION = 2, // Ubicación comercial
  OPERATION_LOCATIONS = 3, // Lugar(es) de operación
  CONTACT_INFORMATION = 4, // Información de contacto
  PAYMENT_CONDITIONS = 5, // Condiciones de pago
  TAX_CONDITIONS = 6, // Condiciones tributarias

  SERVICES = 7, // Servicios
  CONTACTS = 8, // Contactos
  POLICIES = 9, // Políticas
  SUNAT_INFORMATION = 10, // Información SUNAT
  BANK_INFORMATION = 11, // Información bancaria
  ACCOUNTS = 12, // Cuentas contable
  TECHNICAL_SHEET = 13, // Ficha técnica
  LIST_UNITS = 14, // Listado de unidades
  FREQUENCY = 15, // Frecuencia
  ROUTES_STATIONS = 16, // Rutas y estaciones
  CONFIGURATION = 17, // Configuración
}